Move relevant code from api and activities to new exports file
